refactor: change the structure of graphql plugin
